chore: update project structure;
