+++ docs/idling.gif
